removed register component but fixed passport session issue after register
