+++ src/components/Authentication.js
@@ -1,9 +1,9 @@
 import React, { Component } from "react";
-import { register } from "../store";
+import { login, register } from "../store";
 import { connect } from "react-redux";
 import { Button, Container } from "@material-ui/core";
 
-class Login extends Component {
+class Authentication extends Component {
   constructor() {
     super();
     this.state = {
@@ -14,6 +14,7 @@ class Login extends Component {
       data: "",
     };
     this.handleChange = this.handleChange.bind(this);
+    this.login = this.login.bind(this);
     this.register = this.register.bind(this);
   }
 
@@ -32,6 +33,12 @@ class Login extends Component {
     });
   }
 
+  async login(event) {
+    event.preventDefault();
+    await this.props.dispatchUser(this.state.username, this.state.password);
+    this.props.history.push("/");
+  }
+
   async register(event) {
     event.preventDefault();
     await this.props.dispatchNewUser(
@@ -40,15 +47,52 @@ class Login extends Component {
       this.state.email,
       this.state.zipCode
     );
+    await this.props.dispatchUser(this.state.username, this.state.password);
     this.props.history.push("/");
   }
 
   render() {
     return (
       <Container maxWidth="sm">
+        <form onSubmit={this.login}>
+          <h1 className="textCenter" style={{ paddingLeft: "210px" }}>
+            Login
+          </h1>
+          <span style={{ paddingLeft: "100px" }}>
+            <input
+              placeholder="username"
+              name="username"
+              onChange={this.handleChange}
+              required
+            />
+            <span style={{ paddingLeft: "5px" }}>
+              <input
+                placeholder="password"
+                type="password"
+                name="password"
+                onChange={this.handleChange}
+                required
+              />
+            </span>
+            <span style={{ paddingLeft: "10px" }}>
+              <Button
+                type="submit"
+                color="inherit"
+                style={{
+                  color: "white",
+                  backgroundColor: "#00e600",
+                }}
+              >
+                Submit
+              </Button>
+            </span>
+          </span>
+        </form>
+        <br></br>
+        <br></br>
         <form onSubmit={this.register}>
           <h1 className="textCenter" style={{ paddingLeft: "200px" }}>
-            Register
+            Or Sign Up!
           </h1>
           <div style={{ paddingLeft: "100px" }}>
             <input
@@ -109,8 +153,9 @@ const mapState = (state) => ({
 });
 
 const mapDispatch = (dispatch) => ({
+  dispatchUser: (username, password) => dispatch(login(username, password)),
   dispatchNewUser: (username, password, email, zipCode) =>
     dispatch(register(username, password, email, zipCode)),
 });
 
-export default connect(mapState, mapDispatch)(Login);
+export default connect(mapState, mapDispatch)(Authentication);
