+++ src/components/index.js
@@ -1,4 +1,3 @@
 export { default as NavBar } from "./NavBar";
-export { default as Login } from "./Login";
+export { default as Authentication } from "./Authentication";
 export { default as Home } from "./Home";
-export { default as Register } from "./Register";
